# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "Guardian"
  description: "The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation."
  version: "2.9.0"
  contact:
    name: "API developer"
    url: "https://envisionblockchain.com"
    email: "info@envisionblockchain.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
- url: /api/v1
  description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    Contract:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        description:
          type: string
        owner:
          type: string
        isOwnerCreator:
          type: string
        status:
          type: string
    RetireRequest:
      type: object
      properties:
        id:
          type: string
        contractId:
          type: string
        baseTokenId:
          type: string
        owner:
          type: string
        oppositeTokenId:
          type: string
        baseTokenCount:
          type: number
        oppositeTokenCount:
          type: number
...
paths:
...
  /contracts:
    get:
      tags:
        - contracts
      description: Returns all contracts.
      security:
      - bearerAuth: []
      summary: Returns all contracts.
      parameters:
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: The number of pages to skip before starting to collect the result set
          examples:
            pageIndex:
              summary: Example of a pageIndex
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return
          examples:
            pageSize:
              summary: Example of a pageSize
              value: 100
      responses:
        200:
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contract'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - contracts
      description: Creates new contract. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Create new contract.
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/import:
    post:
      tags:
        - contracts
      description: Creates new contract. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Import new contract.
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                contractId:
                  type: string
                description:
                  type: string
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{contractId}/user:
    post:
      tags:
        - contracts
      description: Add new contract user. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Add new contract user.
      parameters:
        - in: path
          name: contractId
          schema:
            type: string
          description: Contract identifier
          required: true
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{contractId}/status:
    post:
      tags:
        - contracts
      description: Update contract status. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Update contract status.
      parameters:
        - in: path
          name: contractId
          schema:
            type: string
          description: Contract identifier
          required: true
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{contractId}/pair:
    post:
      tags:
        - contracts
      description: Creates new contract pair. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Creates new contract pair.
      parameters:
        - in: path
          name: contractId
          schema:
            type: string
          description: Contract identifier
          required: true
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                baseTokenId:
                  type: string
                oppositeTokenId:
                  type: string
                baseTokenCount:
                  type: string
                oppositeTokenCount:
                  type: string
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/pair:
    get:
      tags:
        - contracts
      description: Returns all contracts pairs.
      security:
      - bearerAuth: []
      summary: Returns all contracts pairs.
      parameters:
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: The number of pages to skip before starting to collect the result set
          examples:
            pageIndex:
              summary: Example of a pageIndex
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return
          examples:
            pageSize:
              summary: Example of a pageSize
              value: 100
      responses:
        200:
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    baseTokenRate:
                      type: string
                    oppositeTokenRate:
                      type: string
                    contractId:
                      type: string
                    description:
                      type: string
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/{contractId}/retire/request:
    post:
      tags:
        - contracts
      description: Creates new contract retire request.
      security:
      - bearerAuth: []
      summary: Creates new contract retire request.
      parameters:
        - in: path
          name: contractId
          schema:
            type: string
          description: Contract identifier
          required: true
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                baseTokenId:
                  type: string
                oppositeTokenId:
                  type: string
                baseTokenCount:
                  type: string
                oppositeTokenCount:
                  type: string
                baseTokenSerials:
                  type: string
                oppositeTokenSerials:
                  type: string
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetireRequest"
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/retire/request:
    get:
      tags:
        - contracts
      description: Returns all contract requests.
      security:
      - bearerAuth: []
      summary: Returns all contract requests.
      parameters:
        - in: query
          name: contractId
          schema:
            type: string
          description: Contract Identifier
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: The number of pages to skip before starting to collect the result set
          examples:
            pageIndex:
              summary: Example of a pageIndex
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return
          examples:
            pageSize:
              summary: Example of a pageSize
              value: 100
      responses:
        200:
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RetireRequest"
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - contracts
      description: Cancel contract requests.
      security:
      - bearerAuth: []
      summary: Cancel contract requests.
      parameters:
        - in: query
          name: requestId
          schema:
            type: string
          description: Request Identifier
          required: true
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /contracts/retire:
    post:
      tags:
        - contracts
      description: Retire tokens. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Retire tokens.
      requestBody:
        description: Request Object Parameters.
        content:
          application/json:
            schema:
              type: object
              properties:
                requestId:
                  type: string
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'